rename swerve to bot spefii name
